+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-06T05:31:53.389Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-06T11:09:54.594Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")