+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-06T05:31:56.814Z")
+    created: Optional[datetime] = parse_datetime("2026-02-06T11:09:57.869Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")