+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGRPZAETZ325F59S0K9CTB6P"
+        "01KGSAA7ZY9NBAKZ9EP1BFFNPG"
     )
     r"""The unique identifier of the data source"""