+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGRPZAEXRJW0KWMRWY45S6TM"
+        "01KGSAA800Y92B9WSXXJP0ZF78"
     )
     r"""The unique identifier of the data source"""